etl-final/metadata-service/requirements.txt
Django
djangorestframework
requests
kafka-python
